refactor components into own submodules
