+++ components/project-list/item-delegate.go
@@ -1,4 +1,4 @@
-package components
+package projectlist
 
 import (
 	"fmt"
