+++ components/project-list/msg.go
@@ -1,0 +1,8 @@
+package projectlist
+
+import "github.com/charmbracelet/bubbles/list"
+
+type fatalErrorMsg struct {
+	err error
+}
+type initMsg struct{ items []list.Item }
