+++ components/search-input/msg.go
@@ -1,0 +1,9 @@
+package searchinput
+
+type CancelSearch struct{}
+type UpdateSearch struct {
+	FilteredItemsIndices []int
+}
+type SubmitSearch struct {
+	FilteredItemsIndices []int
+}
