+++ components/styles/util.go
@@ -2,7 +2,7 @@ package styles
 
 import "github.com/charmbracelet/lipgloss"
 
-func baseTitle() lipgloss.Style {
+func BaseTitle() lipgloss.Style {
 	return lipgloss.NewStyle().
 		Bold(true).
 		Foreground(lipgloss.Color("#FFFDF5")).
